+++ react-app/src/components/form/TextFormField.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import { TextField, TextFieldProps } from '@mui/material';
 import { useFormContext } from 'react-hook-form';
 
-const TextInput = (props: TextFieldProps) => {
+const TextFormField = (props: TextFieldProps) => {
   const { register } = useFormContext();
   const { name } = props;
   return <TextField {...props} {...register(name)} />;
 };
 
-export default TextInput;
+export default TextFormField;
